+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
-##
+## 0.3.9
 - Add support for deprecated version of `LAN_RAILCOM_GETDATA` ([#13](https://github.com/ZIMO-Elektronik/Z21/issues/13))
+- Bugfix broadcast `LAN_SYSTEMSTATE_DATACHANGED` automatically when detecting changes ([#28](https://github.com/ZIMO-Elektronik/Z21/issues/28))
 
 ## 0.3.8
 - Add undocumented CV commands ([#25](https://github.com/ZIMO-Elektronik/Z21/issues/25))